feat(solar-simulator): make deploy-mock-readings deploy for each asset in a new worker process
